requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
prometheus-client==0.20.0
pydantic==2.7.4
PyYAML==6.0.1
sqlite-utils==3.36
anyio==4.4.0
pytest==8.2.2
httpx==0.27.0
pytest-asyncio==0.23.7
python-dotenv==1.0.1
ruff==0.5.5
mypy==1.10.0
groq==0.9.0
